+++ packages/agent/src/index.ts
@@ -5,7 +5,6 @@ export type * from './types/identity-vault.js';
 export type * from './types/key-manager.js';
 export type * from './types/sync.js';
 export type * from './types/vc.js';
-export type * from './prototyping/clients/web5-rpc-types.js';
 
 export * from './bearer-identity.js';
 export * from './crypto-api.js';
@@ -13,9 +12,8 @@ export * from './did-api.js';
 export * from './dwn-api.js';
 export * from './hd-identity-vault.js';
 export * from './identity-api.js';
-export * from './prototyping/clients/json-rpc.js';
 export * from './local-key-manager.js';
-export * from './prototyping/clients/web5-rpc-client.js';
+export * from './rpc-client.js';
 export * from './store-data.js';
 export * from './store-did.js';
 export * from './store-identity.js';
