+++ packages/agent/src/test-harness.ts
@@ -11,7 +11,7 @@ import type { Web5PlatformAgent } from './types/agent.js';
 import { AgentDidApi } from './did-api.js';
 import { AgentDwnApi } from './dwn-api.js';
 import { AgentSyncApi } from './sync-api.js';
-import { Web5RpcClient } from './prototyping/clients/web5-rpc-client.js';
+import { Web5RpcClient } from './rpc-client.js';
 import { AgentCryptoApi } from './crypto-api.js';
 import { AgentIdentityApi } from './identity-api.js';
 import { BearerIdentity } from './bearer-identity.js';
